refactor(application): lineService에서 구간에 대한 역할을 수행 하도록 변경
