+++ src/test/java/wooteco/subway/application/LineServiceTest.java
@@ -17,7 +17,6 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.SectionDao;
-import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Section;
 import wooteco.subway.dto.LineResponse;
@@ -31,9 +30,9 @@ class LineServiceTest {
     @Mock
     private LineDao lineDao;
     @Mock
-    private StationDao stationDao;
-    @Mock
     private SectionDao sectionDao;
+    @Mock
+    private StationService stationService;
 
     @Test
     @DisplayName("중복되지 않은 이름의 노선을 저장")
@@ -44,8 +43,8 @@ class LineServiceTest {
         final Line line = Line.initialCreateWithId(1L, name, color, 강남역, 청계산입구역, 1);
         given(lineDao.existsByName(name)).willReturn(false);
         given(lineDao.save(any())).willReturn(line);
-        doReturn(강남역).when(stationDao).findById(1L);
-        doReturn(청계산입구역).when(stationDao).findById(2L);
+        doReturn(강남역).when(stationService).findStationById(1L);
+        doReturn(청계산입구역).when(stationService).findStationById(2L);
 
         //when
         final LineResponse lineResponse = lineService.save(name, color, 1L, 2L, 1);
